renamed single styled.js files to index.js, removed capitalized letter from animation varibale, moved images to folders where are used
