+++ src/common/Loader/index.js
@@ -3,7 +3,7 @@ import styled, { keyframes } from "styled-components";
 const circleWidth = 12;
 const circleSize = 91;
 
-const Spin = keyframes`
+const spin = keyframes`
     to {
         transform: rotate(360deg);
     }
@@ -16,5 +16,5 @@ export const Loader = styled.div`
     border-radius: 50%;
     border: ${circleWidth}px solid ${({ theme }) => theme.color.gray};
     border-left-color: ${({ theme }) => theme.color.black};
-    animation: ${Spin} 1s linear infinite;
+    animation: ${spin} 1s linear infinite;
 `;
