+++ src/common/SearchResults/index.js
@@ -4,12 +4,12 @@ import { MoviesWrapper } from "../../features/movies/MovieList/styled";
 import { PeopleWrapper } from "../../features/people/PeopleList/styled";
 import { selectMoviesFetchStatus, selectMoviesList, selectMoviesPageCount, selectMoviesResultCount } from "../../features/movies/moviesSlice";
 import { ErrorPage } from "../ErrorPage";
-import { Loader } from "../Loader/styled";
+import { Loader } from "../Loader";
 import { Pagination } from "../Pagination";
 import { useQueryParameter } from "../useQueryParameter";
 import { StyledNoResults } from "./styled";
-import { MainWrapper } from "../MainWrapper/styled";
-import { Header } from "../Header/styled";
+import { MainWrapper } from "../MainWrapper";
+import { Header } from "../Header";
 import { selectPeopleFetchStatus, selectPeopleList, selectPeopleResultCount, selectPeoplePageCount } from "../../features/people/peopleSlice";
 import { useLocation } from "react-router-dom";
 import { PersonTile } from "../PersonTile";
