+++ src/common/SearchResults/styled.js
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { ReactComponent as NoResults } from "../../images/noresults.svg";
+import { ReactComponent as NoResults } from "./noresults.svg";
 
 export const StyledNoResults = styled(NoResults)`
     width: 668px;
